make fetch happen (#167)
* add provisional NEWS item

* clean up testing

* start to build out image tests

* add new tests for images

* hot-wire testing

* build fetch and load alternatives

* return value for glue_js()

* test fetch and load using V8

* adjust test for windows

* add reference

* something approximating progress - new fn vw_to_svg_new()

* fetch happens 🎉, need to promote vw_to_svg_new()

* see if we get the same csv data

* take care of timezone issue

* doh!

* try again

* add json handling, remove datetime variables from tests

* remove negative values

* promote vw_to_svg_new() to vw_to_svg()

* replace node references with V8

* update dependencies
