+++ .github/workflows/R-CMD-check.yaml
@@ -9,6 +9,7 @@ on:
     branches:
       - main
       - master
+      - update-159-vegalite-5
 
 name: R-CMD-check
