+++ NAMESPACE
@@ -32,6 +32,7 @@ export(vw_add_signal_listener)
 export(vw_as_json)
 export(vw_autosize)
 export(vw_examine)
+export(vw_fetch)
 export(vw_handler)
 export(vw_handler_add_effect)
 export(vw_handler_body_compose)
@@ -39,6 +40,7 @@ export(vw_handler_compose)
 export(vw_handler_data)
 export(vw_handler_event)
 export(vw_handler_signal)
+export(vw_load)
 export(vw_rename_datasets)
 export(vw_serialize_data)
 export(vw_set_base_url)
