+++ R/spec-version.R
@@ -14,10 +14,8 @@
 #' @return `list` with elements `library`, `version`
 #' @examples
 #'   vw_spec_version(spec_mtcars)
-#' \dontrun{
-#'   # requires nodejs to be installed
+#'   # vw_to_vega() requires the V8 package
 #'   vw_spec_version(vw_to_vega(spec_mtcars))
-#' }
 #' @export
 #'
 vw_spec_version <- function(spec) {
