+++ man/vw_fetch.Rd
@@ -1,0 +1,29 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/callback-helpers.R
+\name{vw_fetch}
+\alias{vw_fetch}
+\alias{vw_load}
+\title{Callback helpers}
+\usage{
+vw_fetch(url, options = NULL)
+
+vw_load(filename)
+}
+\arguments{
+\item{url}{\code{character}}
+
+\item{options}{named \code{list}, not yet implemented}
+
+\item{filename}{\code{character}}
+}
+\value{
+\code{character} contents of the file or URL
+}
+\description{
+These are used by V8 so that R can access the file-system and the network
+so that external data can be accessed.
+}
+\details{
+Assumes the encoding is UTF-8.
+}
+\keyword{internal}
