+++ man/vw_spec_version.Rd
@@ -25,8 +25,6 @@ Returns a list with two elements:
 }
 \examples{
   vw_spec_version(spec_mtcars)
-\dontrun{
-  # requires nodejs to be installed
+  # vw_to_vega() requires the V8 package
   vw_spec_version(vw_to_vega(spec_mtcars))
-}
 }
